(x86sim) fixed compile problems for test inputs
* Code quality for the tests was not great. I fixed
   - removed unused variables
   - added missing system headers for implicit functions
   - fixed numerous printf-style format strings
   - removed <selinux/selinux.h> dependencies
   - fixed incorrect logic in a couple places
   - fixed a stack array overflow in two places
   - fixed some questionable casting
   - moved declarations for conditionally compiled code
   - added feature declarations for non-standard functions
   - added forward declarations for some test support functions
   (Thanks to Cory Cohen for a few of these)

* Improved the makefile
   - Builds a library for test support
   - Suppress command output for libipc
   - Improved compiling speed for test inputs

+++ projects/simulator/tests/syscall_tst.117.msgctl.02.c
@@ -108,14 +108,14 @@ int main(int ac, char **av)
 
 				if (qs_buf.msg_qbytes == new_bytes) {
 					tst_resm(TPASS, "qs_buf.msg_qbytes is"
-						 " the new value - %d",
-						 qs_buf.msg_qbytes);
+						 " the new value - %ld",
+						 (long)qs_buf.msg_qbytes);
 				} else {
 					tst_resm(TFAIL, "qs_buf.msg_qbytes "
 						 "value is not expected");
-					tst_resm(TINFO, "expected - %d, "
-						 "received - %d", new_bytes,
-						 qs_buf.msg_qbytes);
+					tst_resm(TINFO, "expected - %lu, "
+						 "received - %ld", new_bytes,
+						 (long)qs_buf.msg_qbytes);
 				}
 			} else {
 				tst_resm(TPASS, "msgctl() call succeeded");
